Use Homebrew’s `/etc/` management for `tideways-php@*`
This prevents the INI file from being overwritten (and the API key deleted)
whenever the Formula is upgraded.

When using `.install` for a file in `/etc/`, Homebrew will create `.default`
files on upgrade and reinstall when the current config differs from the (new)
upstream config, allowing users to easily diff the changes when desired.